Feat : 사장 가게 메뉴 수정

+++ front/src/components/owner/store/StoreList.tsx
@@ -44,14 +44,17 @@ function StoreList() {
   };
 
   const handleDelete = (id: number) => {
-    API.delete(`stores/${id}`, { params: { storeId: id } })
-      .then((response: any) => {
-        console.log(response.data);
-        setIsRegisterStore(!isRegisterStore);
-      })
-      .catch((error: any) => {
-        console.error(error);
-      });
+    const confirmed = window.confirm('정말 삭제하시겠습니까?');
+    if (confirmed) {
+      API.delete(`stores/${id}`, { params: { storeId: id } })
+        .then((response: any) => {
+          console.log(response.data);
+          setIsRegisterStore(!isRegisterStore);
+        })
+        .catch((error: any) => {
+          console.error(error);
+        });
+    }
   };
 
   useEffect(() => {
